Separate files struct and methods to the other go file.
